优化搜索

+++ user-service/user-service-client/src/main/java/com/gitgle/service/resp/SearchResp.java
@@ -11,9 +11,9 @@ public class SearchResp implements Serializable {
 
     private Long totalPage;
 
-    private Integer page;
+    private Long page;
 
-    private Integer pageSize;
+    private Long pageSize;
 
     private List<SearchUser> searchUsers;
